🐛 BO1: LOGIN > Fix bug of infinite loading while redirect user to OrderPage
