+++ src/Components/pages/login/LoginForm.jsx
@@ -14,9 +14,9 @@ export default function LoginForm() {
 
   const handleSubmit = async (event) => {
     event.preventDefault();
-    authenticateUser(username);
+    const userReceived = await authenticateUser(username);
     setUsername("");
-    navigate(`order/${username}`);
+    navigate(`order/${userReceived.username}`);
   };
 
   const handleChange = (event) => {
