--- src/api/user.js
@@ -1,31 +1,0 @@
-import { doc, getDoc, setDoc } from "firebase/firestore";
-import { db } from "./firebase-config";
-import { fakeMenu } from "../fakeData/fakeMenu.jsx";
-
-export const getUser = async (userId) => {
-  const docRef = doc(db, "users", userId);
-
-  const docSnapshot = await getDoc(docRef);
-  if (docSnapshot.exists()) {
-    const userReceived = docSnapshot.data();
-    return userReceived;
-  }
-};
-
-export const createUser = (userId) => {
-  const docRef = doc(db, "users", userId);
-
-  const newDoc = {
-    username: userId,
-    menu: fakeMenu.SMALL,
-  };
-
-  setDoc(docRef, newDoc);
-};
-
-export const authenticateUser = async (userId) => {
-  const existingUser = await getUser(userId);
-  if (!existingUser) {
-    createUser(userId);
-  }
-};
